Restructure files, folders
Change-type: patch
