+++ lib/cli/commands/build.js
@@ -10,8 +10,16 @@ const NETLIFY_TOKEN = process.env.NETLIFY_AUTH_TOKEN
 const runner = require('../../build-runner')
 
 const {
-  getCurrentGitReference, getMetaData, log
-} = require('../utils')
+  log
+} = require('../../utils/log')
+
+const {
+  getMetaData
+} = require('../../meta-generator/getMetaData')
+
+const {
+  getCurrentGitReference
+} = require('../../meta-generator/getCurrentGitReference')
 
 class BuildCommand extends Command {
   async run () {
