+++ lib/cli/commands/meta.js
@@ -5,8 +5,16 @@ const fs = require('fs')
 const path = require('path')
 
 const {
-  getCurrentGitReference, getMetaData, log
-} = require('../utils')
+  log
+} = require('../../utils/log')
+
+const {
+  getMetaData
+} = require('../../meta-generator/getMetaData')
+
+const {
+  getCurrentGitReference
+} = require('../../meta-generator/getCurrentGitReference')
 
 class MetaCommand extends Command {
   async run () {
