+++ lib/meta-generator/getCurrentGitReference.js
@@ -1,0 +1,18 @@
+const {
+  log
+} = require('../utils/log')
+
+const git = require('simple-git/promise')()
+
+exports.getCurrentGitReference = async () => {
+  log('gathering git reference information...')
+
+  const branch = await git.revparse([ '--abbrev-ref', 'HEAD' ])
+  const remote = await git.listRemote([ '--get-url', 'origin', '--push' ])
+  const repository = remote.trim().match(/:(.+)\.git$/)[1]
+
+  return {
+    branch,
+    repository
+  }
+}
