+++ lib/meta-generator/getMetaData.js
@@ -1,33 +1,17 @@
-const chalk = require('chalk')
 const scrutinizer = require('scrutinizer')
-const git = require('simple-git/promise')()
-const metaGenerator = require('../../bot/lib/generate-landr-meta')
-
-exports.log = (message) => {
-  console.log(chalk.blue('[landr]'), message)
-}
-
-exports.getCurrentGitReference = async () => {
-  exports.log('gathering git reference information...')
-
-  const branch = await git.revparse([ '--abbrev-ref', 'HEAD' ])
-  const remote = await git.listRemote([ '--get-url', 'origin', '--push' ])
-  const repository = remote.trim().match(/:(.+)\.git$/)[1]
-
-  return {
-    branch,
-    repository
-  }
-}
+const {
+  log
+} = require('../utils/log')
+const metaGenerator = require('./generate-landr-meta')
 
 exports.getMetaData = async (repository, branch, context, downloadRepo) => {
-  exports.log(`generating data for ${repository} on branch ${branch}...`)
+  log(`generating data for ${repository} on branch ${branch}...`)
   let scrutinizerData = await scrutinizer.local(repository, {
     reference: branch,
     downloadRepo,
     context
   })
-  exports.log(`generated data for ${repository} with version ${scrutinizerData.version}...`)
+  log(`generated data for ${repository} with version ${scrutinizerData.version}...`)
 
   if (
     scrutinizerData.orgContract &&
