+++ lib/utils/log.js
@@ -1,0 +1,5 @@
+const chalk = require('chalk')
+
+exports.log = (message) => {
+  console.log(chalk.blue('[landr]'), message)
+}
